moved code into src directory and added inst ructions to readme for creating a bind mount when running container
