+++ src/Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.7.3-alpine
 
-WORKDIR /usr/src/app
+WORKDIR /meep/api/src
 
 COPY requirements.txt .
 
@@ -9,5 +9,6 @@ RUN pip install --upgrade pip \
     && pip install gunicorn==19.9.0
 
 COPY . .
+EXPOSE 8000
 
-CMD ["gunicorn", "--bind", "0.0.0.0:8000", "wsgi"]
+CMD ["gunicorn", "--bind", "0.0.0.0:8000", "--reload", "wsgi"]
